close rollback recovery review gaps
