+++ doc/game/gameplay/gameplay-longrun-p0-replay-rollback-runbook-2026-03-06.md
@@ -19,7 +19,7 @@
 6. 回滚与 replay：通过 Viewer 提交完整 `RollbackV2` 请求。Viewer 解析 C→T 的 finalized journal suffix，先校验 epoch、事件数/序列化字节上限和 runtime canonical journal commitment，再调用 direct runtime `rollback_to_snapshot_with_reconciliation(snapshot_C, journal_through_T, reason, target_batch_id_T, approval, now_ms)`。runtime 继续作为 registry、签名、有效期、nonce、journal commitment、target root 与 reconciliation 的最终判定方；只有 receipt 可序列化且所有校验成功后才能提交 Viewer 的 world、batch、challenge、checkpoint 与 reorg epoch 候选状态。
 7. 恢复对账：再次执行 `first_tick_consensus_drift()` 与 `verify_tick_consensus_chain()`，必须均为“无漂移”；实际 target state root 与 journal commitment 必须与签名 intent/持久化 outcome 一致，committed reorg epoch 必须与 receipt/outcome 一致。
 8. 幂等回执与审计：成功 outcome 按 nonce 与 canonical intent digest 持久化。相同 nonce + 相同 canonical intent 的重试不得再次 mutation 或增加 reorg epoch，必须返回同一逻辑 receipt；相同 nonce + 不同 intent 返回 `rollback_nonce_conflict`。断线重连或 operator 查询使用 `authoritative_recovery.get_rollback_receipt`，不得通过再次执行 rollback 猜测结果。确认 `RollbackApplied`、ticket、两个 authority id、nonce、target commitment/root 与 receipt evidence 均已归档。
-9. 玩家动作处置：对受 C→T/fork 影响的每个动作记录 `preserved_at_target`、`replayed`、`rejected_fork` 或 `compensation_required`。生产路径中，被 fork 拒绝且可归属玩家的动作必须创建 `compensation_required`，关联公开安全的 owner/ticket 别名与状态；状态推进或未知归属修复必须由 active governance rollback authority 对绑定 `issued_at_ms`、`expires_at_ms` 与一次性 nonce 的 canonical 请求签名，并通过 Viewer 与 nonce 消费记录原子持久化。过期、超出 future-skew 或重放请求必须在 mutation 前拒绝。`rejected` 是已裁决终态，不再阻塞 all-clear；缺失玩家归属必须以 batch/event/action 身份进入 census 并留下 `player_attribution_incomplete` blocker，直至授权修复。
+9. 玩家动作处置：对受 C→T/fork 影响的每个动作记录 `preserved_at_target`、`replayed`、`rejected_fork` 或 `compensation_required`。生产路径中，被 fork 拒绝且可归属玩家的动作必须创建 `compensation_required`，关联公开安全的 owner/ticket 别名与状态；状态推进或未知归属修复必须由 active governance rollback authority 对绑定 batch/event、`issued_at_ms`、`expires_at_ms` 与一次性 nonce 的 canonical 请求签名，并通过 Viewer 与 nonce 消费记录原子持久化。过期、超出 future-skew、篡改或重放请求必须在 mutation 前拒绝。`rejected` 是已裁决终态，不再阻塞 all-clear；缺失玩家归属必须以 batch/event/action 身份进入 census 并留下 `player_attribution_incomplete` blocker，直至授权修复。经治理签名确认确属系统自动化而非玩家动作时，可持久化为 `system_authored`：必须保留原 batch/event/action census 证据，不得伪造 player id、玩家补偿或补偿工单；只有该分类与 nonce 原子持久化后，才能移除该条归因 blocker。对玩家/社区只能表述为“已核验的系统自动化/非玩家动作”，不得称为受影响玩家或暗示补偿资格。
 
 ## 3. 演练命令（required-tier）
 ```bash
